Make code organization match the functional structure (#139)
part of https://github.com/datafusion-contrib/datafusion-tui/issues/132

We are moving `dft` so there is an engine and 2 uis (the terminal tui
and the cli)

Let's make the code structure match this functional structure:
1. Rename `ui` to `tui` to make it clear it is the terminal ui
2. Move the CLI related code into its own module (cli)
3. Move the command line parsing structures into `DftArgs` in in main.rs


I don't fully understand the split between what logic belongs in `app`
and what belongs in `tui` but I figure we can work on that split over
time
