+++ src/app/handlers/mod.rs
@@ -33,12 +33,9 @@ use std::sync::Arc;
 #[cfg(feature = "flightsql")]
 use tonic::transport::Channel;
 
-use crate::{
-    app::{state::tabs::history::HistoryQuery, AppEvent},
-    ui::SelectedTab,
-};
-
 use super::App;
+use crate::app::ui::SelectedTab;
+use crate::app::{state::tabs::history::HistoryQuery, AppEvent};
 
 pub fn crossterm_event_handler(event: event::Event) -> Option<AppEvent> {
     match event {
