+++ src/app/ui/tabs/logs.rs
